Changed license files to GPL
We could really release this under a dual license model,
GPL if the user enable the COSMOS-based PSF models and Apache 2.0
if the user doesn't. I'm not sure of the best way to do this,
so we'll just stick with GPL for now. No one is likely to use
this project without using the PSF models, and the more general
purpose filters could potentially be added to ITK in the future.